requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
et-xmlfile==1.1.0
Flask==2.2.3
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.0
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
pandas-stubs==1.5.3.230214
pathspec==0.11.0
platformdirs==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
tomli==2.0.1
tqdm==4.64.1
types-pytz==2022.7.1.2
typing-extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
